Deprecate `BaseProfiler` in favor of `Profiler` (#12150)
