+++ docs/source/api_references.rst
@@ -262,10 +262,9 @@ Profiler API
     :nosignatures:
     :template: classtemplate.rst
 
-    AbstractProfiler
     AdvancedProfiler
-    BaseProfiler
     PassThroughProfiler
+    Profiler
     PyTorchProfiler
     SimpleProfiler
     XLAProfiler
